requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0.1
django-braces==1.15.0
sqlparse==0.4.2
tzdata==2021.5
whitenoise==5.3.0
